backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
playwright
apscheduler
pydantic
python-dotenv
httpx
sqlalchemy
